WBO/WBO-collection/Lab1/requirements.txt
biopython
requests
pylab-sdk
matplotlib
jupyter